Update PersonPhotoCell code implementation from feedback
