+++ Say Their Names/Source/Utils/Extentions/UICollectionViewCell++.swift
@@ -1,0 +1,12 @@
+//
+//  UICollectionViewCell++.swift
+//  Say Their Names
+//
+//  Created by Leonardo Garcia  on 02/06/20.
+//  Copyright © 2020 Franck-Stephane Ndame Mpouli. All rights reserved.
+//
+
+import UIKit
+
+// MARK: - UICollectionViewCell conforming to `Reusable` protocol.
+extension UICollectionViewCell: Reusable { }
